tweak(Filesystem/Previews): run stat before rmdir
dir (or parent) could have already been removed

Reviewed-on: http://gerrit.tine20.com/customers/12215
Tested-by: Philipp Schüle <p.schuele@metaways.de>

+++ tine20/Tinebase/FileSystem/Previews.php
@@ -238,7 +238,9 @@ class Tinebase_FileSystem_Previews
         if (!$fileSystem->isDir($basePath)) {
             $fileSystem->mkdir($basePath);
         } else {
-            $fileSystem->rmdir($basePath, true);
+            if ($fileSystem->fileExists($basePath)) {
+                $fileSystem->rmdir($basePath, true);
+            }
             $fileSystem->mkdir($basePath);
         }
         $transactionId = Tinebase_TransactionManager::getInstance()->startTransaction(Tinebase_Core::getDb());
